splitting out template commands into separate module
